requirements.txt
python-dotenv
openai
pinecone